+++ Dracula/Artifacts/Duo/DraculaDizzyArtifact.cs
@@ -1,12 +1,14 @@
 ﻿using Nanoray.PluginManager;
 using Nickel;
+using Shockah.Shared;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 
 namespace Shockah.Dracula;
 
-internal sealed class DraculaDizzyArtifact : Artifact, IRegisterable
+internal sealed class DraculaDizzyArtifact : Artifact, IRegisterable, IStatusLogicHook, IHookPriority
 {
 	internal const int ResultingOxidation = 2;
 
@@ -36,4 +38,24 @@ internal sealed class DraculaDizzyArtifact : Artifact, IRegisterable
 			.Concat(StatusMeta.GetTooltips(ModEntry.Instance.OxidationStatus.Status, ResultingOxidation))
 			.Concat(StatusMeta.GetTooltips(Status.corrode, 1))
 			.ToList();
+
+	public double HookPriority
+		=> 1;
+
+	public bool HandleStatusTurnAutoStep(State state, Combat combat, StatusTurnTriggerTiming timing, Ship ship, Status status, ref int amount, ref StatusTurnAutoStepSetStrategy setStrategy)
+	{
+		if (status != ModEntry.Instance.OxidationStatus.Status)
+			return false;
+		if (timing != StatusTurnTriggerTiming.TurnEnd)
+			return false;
+
+		var maxTriggers = state.EnumerateAllArtifacts().Any(a => a is ThinBloodArtifact) ? 2 : 1;
+		var triggers = Math.Min(maxTriggers, ship.Get(ModEntry.Instance.BleedingStatus.Status));
+		if (triggers > 0)
+		{
+			amount += triggers * ResultingOxidation;
+			Pulse();
+		}
+		return false;
+	}
 }