+++ Dracula/docs/release-notes.md
@@ -5,6 +5,7 @@
 ## Upcoming release
 
 * Fixed a crash when previewing Bat Form upgrades when currently on the 4th quadrant of the base card.
+* Fixed the Dizzy duo Oxidation vs Corrode vs Bleeding triggers. It still triggers delayed, but doesn't eat up Oxidation while doing so.
 
 ## 1.1.1
 Released 24 April 2024.